+++ leviathan/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-__version__ = "2025.1.18"
+__version__ = "2025.1.21"
 from . import utils
 from . import index